rename foo.go to fooapp.go to be consistent with ksapp, gcpapp
