+++ bootstrap/cmd/kfctl/cmd/root.go
@@ -31,7 +31,6 @@ import (
 	"path"
 	"path/filepath"
 	"plugin"
-	"strings"
 )
 
 func LoadPlatform(platform string, options map[string]interface{}) (kftypes.KfApp, error) {
@@ -47,12 +46,12 @@ func LoadPlatform(platform string, options map[string]interface{}) (kftypes.KfAp
 		// Comment out this section and comment in the line
 		//   return nil, fmt.Errorf("unknown platform %v", platform
 		plugindir := os.Getenv("PLUGINS_ENVIRONMENT")
-		pluginpath := filepath.Join(plugindir, platform+".so")
+		pluginpath := filepath.Join(plugindir, platform+"app.so")
 		p, err := plugin.Open(pluginpath)
 		if err != nil {
 			return nil, fmt.Errorf("could not load plugin %v for platform %v Error %v", pluginpath, platform, err)
 		}
-		symName := "Get" + strings.ToUpper(platform[0:1]) + platform[1:] + "App"
+		symName := "GetKfApp"
 		symbol, symbolErr := p.Lookup(symName)
 		if symbolErr != nil {
 			return nil, fmt.Errorf("could not find symbol %v for platform %v Error %v", symName, platform, symbolErr)
